+++ pangea-sdk/.sdk-ci.yml
@@ -9,10 +9,10 @@ sdk-test-it:
   retry: 2
   variables:
     # Set each service test environment
+    SERVICE_AI_GUARD_ENV: DEV
     SERVICE_AUDIT_ENV: LVE
     SERVICE_AUTHN_ENV: LVE
     SERVICE_AUTHZ_ENV: LVE
-    SERVICE_DATA_GUARD_ENV: LVE
     SERVICE_EMBARGO_ENV: LVE
     SERVICE_FILE_SCAN_ENV: LVE
     SERVICE_IP_INTEL_ENV: LVE
@@ -20,7 +20,7 @@ sdk-test-it:
     SERVICE_URL_INTEL_ENV: LVE
     SERVICE_FILE_INTEL_ENV: LVE
     SERVICE_USER_INTEL_ENV: LVE
-    SERVICE_PROMPT_GUARD_ENV: LVE
+    SERVICE_PROMPT_GUARD_ENV: DEV
     SERVICE_REDACT_ENV: LVE
     SERVICE_SANITIZE_ENV: LVE
     SERVICE_VAULT_ENV: LVE
@@ -47,6 +47,9 @@ sdk-test-it:
 
   parallel:
     matrix:
+      - CLOUD: [AWS]
+        ENV: ${SERVICE_AI_GUARD_ENV}
+        TEST: ai_guard
       - CLOUD: [AWS, GCP]
         ENV: ${SERVICE_AUDIT_ENV}
         TEST: audit
@@ -56,9 +59,6 @@ sdk-test-it:
       - CLOUD: [AWS, GCP]
         ENV: ${SERVICE_AUTHZ_ENV}
         TEST: authz
-      - CLOUD: [AWS]
-        ENV: ${SERVICE_DATA_GUARD_ENV}
-        TEST: data_guard
       - CLOUD: [AWS, GCP]
         ENV: ${SERVICE_EMBARGO_ENV}
         TEST: embargo
